Delete Common.ts
