+++ packages/dds/tree/src/id-compressor/AppendOnlySortedMap.ts
@@ -5,7 +5,8 @@
 
 /* eslint-disable no-bitwise */
 
-import { assert, fail } from "./Common";
+import { assert } from "@fluidframework/common-utils";
+import { fail } from "../util";
 
 /**
  * A map in which entries are always added in key-sorted order.
