+++ packages/dds/tree/src/id-compressor/index.ts
@@ -41,15 +41,6 @@ export {
     ensureSessionUuid,
 } from "./NumericUuid";
 export {
-    assert,
-    assertNotUndefined,
-    fail,
-    Mutable,
-    ClosedMap,
-    getOrCreate,
-    compareStrings,
-} from "./Common";
-export {
     AttributionId,
     StableId,
     CompressedId,
