+++ packages/dds/tree/src/util/index.ts
@@ -33,14 +33,24 @@ export {
 export { StackyIterator } from "./stackyIterator";
 export {
     clone,
+    compareArrays,
+    compareFiniteNumbers,
+    compareFiniteNumbersReversed,
+    compareMaps,
     compareSets,
+    compareStrings,
+    copyPropertyIfDefined,
     fail,
+    getOrCreate,
+    hasAtLeastLength,
     getOrAddEmptyToMap,
     isJsonObject,
     JsonCompatible,
     JsonCompatibleObject,
     JsonCompatibleReadOnly,
     makeArray,
+    Mutable,
     RecursiveReadonly,
+    setPropertyIfDefined,
     unreachableCase,
 } from "./utils";
